requirements.txt
streamlit>=1.26.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-groq>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
unstructured>=0.10.30
unstructured[pdf]>=0.10.30
python-docx>=0.8.11
docx2txt>=0.8
pypdf>=3.17.0
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.0
torch>=2.0.0
transformers>=4.35.0

selenium>=4.10.0
webdriver-manager>=3.8.6